conjugate_encoder

This function encodes categorical variables by fitting a posterior distribution per each category to the target variable y, using a known conjugate-prior. The resulting mean(s) of each posterior distribution per each category are used as the encodings.

conjugate_encoder(
  X_train,
  X_test = NULL,
  y,
  cat_columns,
  prior_params,
  objective = "regression"
)

Arguments

X_train
A tibble representing the training data set containing some categorical features/columns.

X_test
An optional tibble representing the test set, containing some set of categorical features/columns. Default = NULL.

y
A numeric vector or character vector representing the target variable. If the objective is "binary", then the vector should only contain two unique values.

cat_columns
A character vector containing the names of the categorical columns in the tibble that should be encoded.

prior_params
A list with named parameters that specify the prior assumed. For regression, this requires a dictionary with four keys and four values: mu, vega, alpha, beta. All must be real numbers, alpha should be greater than 0, beta and vega should be greater than 0. mu can be negative. For binary classification, this requires a list with two keys and two values: alpha, beta. All must be real numbers and be greater than 0.

objective
A string, either "regression" or "binary" specifying the problem. Default is regression. For regression, only the uniform quantization method is incorporated here for simplicity.

Value

A list containing with processed training and test sets, in which the named categorical columns are replaced with their encodings. For regression, the encoder will add one additional dimension to the original training set since the assumed prior distribution is two dimensional.

Examples

conjugate_encoder(
X_train = mtcars,
y = mtcars$mpg,
cat_columns = c("gear", "carb"),
prior_params = list(mu = 3, vega = 5, alpha = 3, beta = 3),
objective = "regression")

#> Warning: NA's fitted for expected variance. The variance of a single data point does not exist. Make sure columns specified are truly categorical.

#> Joining, by = "gear"

#> Joining, by = "carb"

#> [[1]]
#>     mpg cyl  disp  hp drat    wt  qsec vs am gear_encoded_mean gear_encoded_var
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1          21.62342         80.61494
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1          21.62342         80.61494
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1          21.62342         80.61494
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0          14.33550         30.90619
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0          14.33550         30.90619
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0          14.33550         30.90619
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0          14.33550         30.90619
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0          21.62342         80.61494
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0          21.62342         80.61494
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0          21.62342         80.61494
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0          21.62342         80.61494
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0          14.33550         30.90619
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0          14.33550         30.90619
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0          14.33550         30.90619
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0          14.33550         30.90619
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0          14.33550         30.90619
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0          14.33550         30.90619
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1          21.62342         80.61494
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1          21.62342         80.61494
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1          21.62342         80.61494
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0          14.33550         30.90619
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0          14.33550         30.90619
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0          14.33550         30.90619
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0          14.33550         30.90619
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0          14.33550         30.90619
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1          21.62342         80.61494
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1          18.89622         80.16129
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1          18.89622         80.16129
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1          18.89622         80.16129
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1          18.89622         80.16129
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1          18.89622         80.16129
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1          21.62342         80.61494
#>    carb_encoded_mean carb_encoded_var
#> 1           14.06162         33.41335
#> 2           14.06162         33.41335
#> 3           22.32355         92.14541
#> 4           22.32355         92.14541
#> 5           19.77838         71.99236
#> 6           22.32355         92.14541
#> 7           14.06162         33.41335
#> 8           19.77838         71.99236
#> 9           19.77838         71.99236
#> 10          14.06162         33.41335
#> 11          14.06162         33.41335
#> 12          14.50270         22.40138
#> 13          14.50270         22.40138
#> 14          14.50270         22.40138
#> 15          14.06162         33.41335
#> 16          14.06162         33.41335
#> 17          14.06162         33.41335
#> 18          22.32355         92.14541
#> 19          19.77838         71.99236
#> 20          22.32355         92.14541
#> 21          22.32355         92.14541
#> 22          19.77838         71.99236
#> 23          19.77838         71.99236
#> 24          14.06162         33.41335
#> 25          19.77838         71.99236
#> 26          22.32355         92.14541
#> 27          19.77838         71.99236
#> 28          19.77838         71.99236
#> 29          14.06162         33.41335
#> 30          17.44324               NA
#> 31          13.37838               NA
#> 32          19.77838         71.99236
#>